refactor PinotControllerRestClient
